product_service/requirements.txt
django==5.1
djangorestframework==3.15
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0
